Added insert and update methods to model classes
This commit includes the addition of insert and update methods to several model classes such as Category, Keyword, Issue, User, Comment, IssueKeyword, State, UserRole, KnowledgeBaseArticle. These methods will allow for easier manipulation of data in the database from the various parts of the application. A change has also been made in the getUserByUsername method of the User class, making the _username parameter a String instead of an integer.